+++ src/infos/options.jl
@@ -220,6 +220,8 @@ end
   dcsd_ofac::Float64 = 0.15
   """`⟨false⟩` ignore various errors in sanity checks. """
   ignore_error::Bool = false
+  """`⟨false⟩` keep the orbitals after rotations over iterations of orbital optimizations in the OQV-CCD/DCD."""
+  keepOQVorbitals::Bool = false
 end
 
 """ 
@@ -252,6 +254,12 @@ end
   cartesian::Bool = false
   """`⟨1000⟩` target batch length for the integral transformation. """
   target_batch_length::Int = 1000
+  """`⟨false⟩` use fallback basis sets (in case of missing basis sets). """
+  use_fallback_basis::Bool = false
+  """`⟨true⟩` sanity check of the fit basis (i.e., that it's not an AO basis)"""
+  check_fit_basis::Bool = true
+  """`⟨true⟩` split independent angular shells (important for efficiency). """
+  split_ashells::Bool = true
 end
 
 """