+++ src/system/basisset.jl
@@ -22,7 +22,7 @@ export generate_angularshell, add_subshell!, angularshells
 export n_subshells, n_primitives, n_coefficients, n_angularshells, n_ao
 export normalize_contraction
 export coefficients_1mat, n_coefficients_1mat
-export basis_name, generate_basis, guess_norb
+export basis_name, generate_basis, guess_norb, output_basis, get_available_elements4basis
 export ao_list, print_ao, ao_order2internal
 export subshell_char, max_l
 export levenshtein_distance, get_available_basis_sets, suggest_basis_sets
@@ -143,6 +143,39 @@ function Base.show(io::IO, bs::BasisSet)
   end
 end
 
+
+"""
+    output_basis([io::IO,] basisname, list_of_elements=String[])
+
+  Output the basis set for the specified elements. If the list is empty, 
+  generate the basis for all elements available in the basis set.
+"""
+function output_basis(io::IO, basisname, list_of_elements=String[])
+  if isempty(list_of_elements)
+    list_of_elements = get_available_elements4basis(basisname)
+  end
+  for atom in list_of_elements
+    # Generate a system with one atom
+    ac = ACentre(atom, 0.0, 0.0, 0.0, Dict("ao" => basisname))
+    system = MSystem([ac])
+    # Generate the basis for the atom
+    basis = generate_basis(system; basisset=basisname, split_ashells=false)
+    bc = basis.centres[1]
+    nume4l = number_of_primitives_for_l(bc)
+    numc4l = number_of_contractions_for_l(bc)
+    println(io, "!")
+    # name of the element, 20 characters long
+    name = lowercase(element_fullname(ac))
+    name *=" "^(20-length(name))
+    println(io, "! ", name, " (", gen_ls_string(nume4l), ") -> [", gen_ls_string(numc4l), "]")
+    for shell in bc.shells
+      println(io, shell)
+    end
+  end
+end
+
+output_basis(basisname, list_of_atoms=String[]) = output_basis(stdout, basisname, list_of_atoms)
+
 """
     shell_range(bs::BasisSet, i::Int=1)
 
@@ -193,19 +226,36 @@ end
   If `basisset` is provided, it is used as the basis set.
 """
 function generate_basis(EC::AbstractECInfo, type="ao"; basisset::AbstractString="")
-  return generate_basis(EC.system, type; cartesian=EC.options.int.cartesian, basisset=basisset)
+  return generate_basis(EC.system, type; cartesian=EC.options.int.cartesian, basisset=basisset,
+                        use_fallback=EC.options.int.use_fallback_basis,
+                        check_fit_basis=EC.options.int.check_fit_basis,
+                        split_ashells=EC.options.int.split_ashells)
 end
 
 """
-    generate_basis(ms::MSystem, type="ao"; cartesian=false, basisset::AbstractString="")
+    generate_basis(ms::MSystem, type="ao"; cartesian=false, basisset::AbstractString="", 
+                   use_fallback=false, check_fit_basis=true, split_ashells=true)
 
   Generate basis sets for integral calculations.
 
   The basis set is stored in [`BasisSet`](@ref) object.
   `type` can be `"ao"`, `"mpfit"` or `"jkfit"`.
   If `basisset` is provided, it is used as the basis set.
+  If `split_ashells` is true, independent angular shells are split (important for efficiency).
 """
-function generate_basis(ms::MSystem, type="ao"; cartesian::Bool=false, basisset::AbstractString="")
+function generate_basis(ms::MSystem, type="ao"; cartesian=false, basisset::AbstractString="",
+                        use_fallback=false, check_fit_basis=true, split_ashells=true)
+  fallback = ""
+  if use_fallback
+    # set fallback basis sets
+    if type == "ao"
+      fallback = "def2-tzvppd"
+    elseif type == "mpfit"
+      fallback = "def2-tzvppd-mpfit"
+    elseif type == "jkfit"
+      fallback = "aug-def2-universal-jkfit"
+    end
+  end
   array_of_centres = BasisCentre[]
   id = 1
   for atom in ms
@@ -217,7 +267,8 @@ function generate_basis(ms::MSystem, type="ao"; cartesian::Bool=false, basisset:
         basisname = guess_basis_name(atom, type)
       end
     end
-    basisfunctions = parse_basis(basisname, atom; fallback=(type=="jkfit"))
+    check_fit_basis_name(basisname, type, check_fit_basis)
+    basisfunctions = parse_basis(basisname, atom; fallback, split_ashells)
     id = set_id!(basisfunctions, id)
     push!(array_of_centres, BasisCentre(atom, basisname, basisfunctions))
   end
@@ -347,7 +398,29 @@ function guess_basis_name(atom::ACentre, type)
     error("AO basis set for atom $(atomic_centre_label(atom)) not defined!")
   end
   aobasis = basis_name(atom, "ao")
+  if startswith(aobasis, "{")
+    error("The basis block in $aobasis is given explicitly, cannot guess the corresponding $type basis")
+  end
   return aobasis * "-" * type
+end
+
+"""
+    check_fit_basis_name(basis_name, type, err)
+
+  Check if the basis name is a valid fit basis (to catch a common mistake of using AO basis 
+  as a fitting basis).
+"""
+function check_fit_basis_name(basis_name, type, err)
+  if startswith(basis_name, "{")
+    # the basis block is given explicitly, cannot check 
+    return
+  end
+  if endswith(type, "fit") && !endswith(basis_name, "fit")
+    warn("Basis set $basis_name is not a valid fitting basis! 
+           Use a fitting basis set (e.g., `avtz-jkfit` for JK fit or `avtz-mpfit` for MP fit) 
+           instead of AO basis set!
+           This error can be ignored by setting option `int.check_fit_basis=false`)", err)
+  end
 end
 
 """