+++ CHANGELOG.md
@@ -4,11 +4,24 @@
 
 ### Breaking
 
+* The fallback basis sets are not used by default anymore. Set `@set int use_fallback_basis=true`
+  to enable them.
+
 ### Changed
+
+* The fallback basis sets are now defined for `ao` (`def2-tzvppd`), `jkfit` (`aug-def2-universal-jkfit`), and `mpfit` (`def2-tzvppd-mpfit`) basis sets.
+* Functions for H and He are copied from [aug]-cc-pVXZ basis sets to [aug]-p[w]CVXZ basis sets. Functions for Li and Be are copied from [aug]-cc-pCVXZ basis sets to [aug]-pwCVXZ basis sets.
 
 ### Added
 
+* automatic augmentation of basis sets by additional diffuse or steep functions.
+* functions to get all elements available in a given basis set (`get_available_elements4basis`) and to output the basis set for a given list of elements in the molpro format (`output_basis`).
+* augmented basis sets for jkfit vXz-jkfit and def2-universal-jkfit basis sets.
+
 ### Fixed
+
+* A simple sanity check of the fitting basis sets is performed (by checking whether it's 
+  an AO basis set). The error message can be turned to a warning by setting `@set int check_fit_basis=false`.
 
 ## Version [v0.14.1] - 2025.07.03